fix: :bug: fixes feedback
