+++ app/[lng]/(personal)/page.js
@@ -1,6 +1,6 @@
 // Import your Client Component
 import MainPage from './client'
-import { getUser } from './actions'
+import { getUser } from '@/app/actions/getUser'
 
 export default async function Page({ params: { lng } }) {
   const user = await getUser()
