+++ app/[lng]/(personal)/tokens/actions/createToken.js
@@ -1,7 +1,7 @@
 'use server'
 
 import { supabaseService } from '@/app/supabase/service'
-import { getUser } from '../../actions'
+import { getUser } from '@/app/actions/getUser'
 import { revalidatePath } from 'next/cache'
 
 export async function createToken(formData) {
@@ -32,31 +32,6 @@ export async function createToken(formData) {
     revalidatePath('/[lng]/tokens', 'page')
     return { success: true, token: data[0] }
   } catch (error) {
-    return { success: false, message: error.message }
-  }
-}
-
-export async function deleteToken(tokenName) {
-  try {
-    if (!tokenName) {
-      throw new Error('Token name is required')
-    }
-
-    const user = await getUser()
-    const { data, error } = await supabaseService
-      .from('tokens')
-      .delete()
-      .eq('name', tokenName)
-      .eq('user_id', user.id)
-
-    if (error) {
-      throw error
-    }
-
-    revalidatePath('/[lng]/tokens', 'page')
-    return { success: true, token: data[0] }
-  } catch (error) {
-    console.error('Error deleting token:', error.message)
     return { success: false, message: error.message }
   }
 }
